Fix PowerPC frequency options, frameskip logic, and audio sync (related to #4)
- Remove unrealistic PPC frequency options (8, 16, 25 MHz); keep range
  33-200 MHz with descriptive labels tied to hardware stepping defaults

- Add missing 33 MHz case to the PPC bus multiplier switch in Model3.cpp

- Fix frameskip counter using modulo arithmetic for a consistent render
  pattern (skip N, render 1) instead of the previous jittery increment
  logic

- Add AdjustAudioForCPUFrequency() to libretroAudio.cpp: scales
  samples-per-frame proportionally when the PPC is underclocked, so
  audio stays in sync and avoids crackling at reduced frequencies

- Change resolution_multiplier from int to float in CoreOptionsTypes.h
  to support the new 0.5x (Half, 248x192) option alongside existing
  1x/2x/3x/4x modes; update all dependent code in libretro.cpp

- Add 'Half (248x192)' internal resolution option targeted at low-end
  mobile devices (RPi5, Mali-G52) where GPU load reduction matters

+++ Src/OSD/libretro/CoreOptionsTypes.h
@@ -2,7 +2,7 @@
 #define SUPERMODEL_CORE_OPTIONS_TYPES_H
 
 struct CoreOptions {
-   int resolution_multiplier;
+   float resolution_multiplier;
    bool widescreen;
    bool vsync;
    bool crosshairs;
